Initial support for TLS communication, for use in Tor.
This adds a basic library for implementing TLS in Haskell. It has a lot of
holes. For example, it only supports RSA with AES, and even then in a limited
number of key exchange modes. At some later date, this work should be expanded
to cover more conenction cases and cipher suites, and then broken out into its
own library. For now, it appears to pass a few simple test cases, and is
suitable for starting the actual Tor implementation.

(This merge is a --squash merge of the initial branch, if you're looking for
more history.)