Node will now automatically download router list consensus documents.
This adds the ability for a Tor node to download, decompress, parse, and
validate consensus documents. Included in this work is the ability to
automatically select a random source for router information, download keys for
the initial server list, and cross-check all signatures used in the various
downloads and file formats.

(This is a squashed commit of the 'inflate' and 'tordir' branches.)