Support for creating multi-link Tor circuits.
This adds the ability to create multi-node circuits, with each step negotiated
using the Tor protocol layers. Currently, only CREATE and EXTEND methods are
supported. Future versions should be extended to support CREATE2 and EXTEND2,
but this requires some cryptography not currently available, at least in pure
Haskell.

This update also provides some modifications to the TLS protocol, to allow the
ability to write to a TLS link while blocked reading from it.

(This is a squashed merge of the "circuit" branch.)