Convert to using Vincent Hanquez's cryptonite and tls libraries, which appear to now fully support the requirements of Tor and have better community support.
The obvious consequence of this merge is the removal of the zlib and TLS
libraries from the source tree, as they are now implemented as external
libraries. In addition, cryptonite and friends tend to utilize strict
ByteStrings instead of lazy ByteStrings, which requires a great deal of
intervention in some of the core parsing libraries. (FIXME: Future commits
should streamline the current parsing system into a more consistent strict
receive / lazy send pattern.)

This commit should be functionally equivalent to the prior commit, but with huge
changes to the underlying structure.

(This is a squash merge of the "cryptonite" branch.)