Remove STM from the system.
This commit removes STM from the system, which simplifies quite a bit
of code, although it introduces the possibility for deadlock in the
system. Where obvious, this also simplifies the code that was supporting
STM; for example, those situations where an IO action was returned from
an STM action.

(This is a squash merge of the "de-stm" branch.)