Basic HaLVM support.
This commit modifies the system to properly support execution on the
HaLVM. Users should be able to use the included demonstration to create
a Tor-anonymized connection to an HTTP host of their choice. It is
probably that further work will be required in the future, in order to
more properly handle relay and exit nodes.

(This is a squash merge of the "halvm" branch.)