Refactoring into a cleaner high-level interface.
This commit updates the code to use a much cleaner separation of concerns,
and provides a nice, simple high-level interface to its use. Functionality
should be the same before and after the commit; in other words, it should
still function as an entrance node, with little in the way of relay or
exit capabilities. These should come in the next set of commits. This code
should support GHC 7.8 and 7.10 series.

(This is a squash merge of the "highlevel" branch.)