Add support for the NTor handshake and relaying.
This commit updates the code to support both the new Tor handshake using
Curve25519, as well as to support relay and exit nodes. Note that this code is
not nearly as well tested as some previous commits, and should probably not be
expected to work all that well. More investigation should be done to confirm
that the mechanisms inserted work properly.

(This is a squash merge of the "relay" branch.)