Improve the Haddock and README documentation.
This patch adds Haddock docs to most of the exported variables in the project,
along with summaries for some of the more important modules. Better
documentation would probably be better, but I think this is a decent base. I've
also updated the README and LICENSE to be accurate, and tested the HaLVM build
instructions against a fairly vanilla HaLVM install.

(This is a squash merge of the "docs" branch.)