Additional small changes to support better distribution and wider library support.
This commit makes sure that all the test files get appropriately picked up when
packaged with cabal sdist. In addition, it modifies the Travis build to try to
make sure that we don't keep having this problem. Finally, at a user's request,
we loosen some of the requirements, and allow newer versions of cryptonite and
memory. The former requires some CPP ugliness, which we will need to deprecate
sometime in the future.

(This is a squashed merge of the "bug" branch.)